+++ packages/lib-classifier/src/components/Classifier/components/TaskArea/components/Tasks/components/TaskNavButtons/components/DoneButton/DoneButton.js
@@ -1,6 +1,5 @@
-import { withThemeContext } from '@zooniverse/react-components'
+import { PrimaryButton, withThemeContext } from '@zooniverse/react-components'
 import counterpart from 'counterpart'
-import { Button, Text } from 'grommet'
 import PropTypes from 'prop-types'
 import React from 'react'
 
@@ -8,12 +7,6 @@ import en from './locales/en'
 import theme from './theme'
 
 counterpart.registerTranslations('en', en)
-
-const Label = (
-  <Text size='medium'>
-    {counterpart('DoneButton.done')}
-  </Text>
-)
 
 // TODO add back gold standard and demo buttons using grommet Button icon prop
 // {props.demoMode && <i className="fa fa-trash fa-fw" />}
@@ -27,11 +20,21 @@ function DoneButton ({
 }) {
   if (!completed) {
     return (
+<<<<<<< HEAD
       <Button
         disabled={disabled}
         label={Label}
         onClick={onClick}
         type='submit'
+=======
+      <PrimaryButton
+        color='green'
+        disabled={props.disabled}
+        fill
+        label={counterpart('DoneButton.done')}
+        onClick={props.onClick}
+        style={{ textTransform: 'capitalize' }}
+>>>>>>> Fix regressions and update text color on green button for accessibility
       />
     )
   }